Changed traitor and traitor-chan scaling.
Other misc changes including map changes:
Re-enabled numbered displays on bags to test the font fix.
Astleymov added to AI chamber.
Dressers added to map, for underwear changes.  Missed that tg update previously.
Another tiny hole in the wall added.  This one's more out of the way, and dangerous to get to.  There is a canonical falsewall entrance, or you can break the wall to get there.
Made the captain's gloves a bit more robust.
Fixed packagewrap, so that packages are now the correct w_class when large objects are wrapped.