Fixes grenade assemblies to my satisfaction (for now).
Grenades with custom assemblies are now referred to as bombs or mines; this is a change on my end, not /tg/'s.  The only thing still considered a grenade is the type made with only a wire (introduced upstream with Pete's changes).  As part of this, they received a graphical change; you can now see the trigger mechanism on its face.  They are actually very ugly and I would prefer to do something about that later.

Because they are bombs and not grenades, using them does not activate them; instead, it shows you the interface for the trigger.  You will also no longer be in throw mode.  In accordance with this, the default time on all timers has increased to 10 seconds from 5.

Additionally, all signalers default to not receiving signals, and an option to turn that on or off is present in their interface.  Infrared beams have a rotate beam option in their interface, when part of an assembly.

This is all sort of hackish.  I would like to do better with it later.