Adds holographic clothes for actors on the theater-holodeck.
The holographic actor's clothing spawns in the theater and requires you to remain there while it is used.  It allows you to summon almost any helmet, gloves, shoes, eyewear, mask, and armor, and changes your jumpsuit like a chamelon jumpsuit.  Only one of each type can be summoned at a time, per jumpsuit.  If you leave the holodeck or remove the suit, all the summoned items vanish.

The way this was implemented it might possibly be abused; I will have to revisit it.  However, for now, I believe the fact that they can only be used while on the holodeck should help limit the potential abuses.  Nevertheless, the items are 'real' and functional until lost.