Theatre update
Gives the theatre jumpsuits an inbuilt microphone.  When toggled on, it relays the wearer's normal speech (without ; or :channel) to a channel designated for theatre use (144.1, the lowest normally-reachable channel).  This channel as yet has no encryption keys nor telecoms tie-in.  This is to allow people to view theatre acts remotely, with speech.

Adds two more scenes to the holodeck, the horrible dance floor from the other holodeck and also a rogue version of the AI floor.  Adds a new beating red floor type for the rogue AI.

Gives the fast food box a bag-like sprite and more food variety.
Adjusts the number of people required for cult.