Timer update
Allows timers to be used in addition to signalers, with anything that uses the wire datum (C4, airlocks, the particle accelerator, etc).

Gives timers a repeat option, so that they can send a signal periodically instead of only counting down once.  Raises the lowest time setting from 1 second to 5 seconds.