Less Job Opening For More Divesity
Decreases the number of job openings of certain jobs to a smaller
number, after careful consideration, to create a bigger diversity in the
roles our players to take.

Talked with a few core regulars, they are in favour of this, but I think
I still need more feedback on these changes.

The HoP can still open new positions for latejoiners if they want to
improve a department, and this will have a greater impact now.