Adds an alt-click to list all the contents of a turf in a stat pane.
Alt-clicking anything in a square will put the contents in the status pane in the top right, in a tab labelled .  The contents of that pane can be left or right clicked on as though you were clicking them on the map; this allows you to pick up, manipulate, examine, etc, anything in an overcrowded square, or to access something otherwise covered by a item with a larger sprite.

This only works on adjacent tiles, and moving away from the tile will cancel it.  Only one tile at a time will be listed.