Magic update
Changes the way wizard spellbooks operate.  Instead of carrying around handy appendicies to all magic in the universe, these appendicies stay mostly at the wizard sanctum.  Instead, a complex system of magic items has been added.

Wizards know a smaller number of innate spells (3), but can enchant items, carry scrolls and spellbooks, and generally do their best to tote a magic armoury around.  At this time only innate spells are subject to "wizard must be dressed up" rules; however, enchanted items MUST be held in an active hand to be used.  This means that if the wizard is stunned, they drop the item, possibly losing it to the crew.  This is dangerous, but a crucial part of being a wizard; that blade of disintegration may be terrible in the wrong hands, but it keeps mental slots free to let you blink away.

At this time the only place to enchant items is the wizard sanctum, and as before there is no way back.  You can rip scrolls out of the master spellbooks and take them with you, or stick the scrolls in spellbooks and items, but once you leave, you are committed.

Additionally, a shadow step spell has been created.  You may jump to a selected shadow if your location, and the other, are both dark.

This update contains an additional mapchange, if you jump to specifically this revision you can wipe anything stupid from it.  I am making poor use of git, I am sorry.