Adds rings
Rings are a subtype of glove and may someday have special powers.  As of now, they are decorative only.  They do not block your fingerprints.

You can make coins into rings with a welder; adds a coin press to the mining asteroid, just somewhat out of the way.  Also moves the wizard sanctuary to hyperspace.