Trimmed a bit of the Too Similar objectives for diversity
-Commented out the following: CMO, RD and CE jumpsuits, and CE cartridge
(RD and CMO cartridges were already commented out, this makes it
consistent)

-Freeform module (Astleymov is a dangerous and unique board found by
nanotrasen! And the two were sitting right next to each other, and were
just renames)
-AI core board and Cyborg upload (AI upload and Borg uploads are just
palette swaps that are sitting right next to each other, and AI upload
is enough for an AI related thing)

-Thermite and Chloral: These are average compounds, and Chloral's making
is wonky, since it results in different amount of chemicals.

TELESCIENCE BUG