Minor fixes and tweaks.
Fixes Adjacent() for diagonals.
Soda cans may now start empty and crushed.
Ghosts will no longer auto-examine turfs on click, although they can still shift-click examine as normal.

Adds a small second EVA to the map with supplies for repairing the bomb range.  It is connected with a glass-wall bridge to science maintenance.  It is AI-monitored and requires engineering access to enter, with EVA access required to get to the space suits.