Tweaks on feedback
*Made reagent objectives Wizard only, as they created weird and awkward
storylines, because nor their absence is noticed, nor does it create
problems when they are found. At least with wizards, it forces them to
spend some time in a location.
*Removed balloon objectives. It was rude, and was already removed.
*Trimmed construction board objectives again, leaving in the AI core
one. Added scientist and roboticist to the banlist, because they have
access to imprinters, and the board requires only glass and sacid.
*Adjusted difficulties to the standards (3: steal something
reproduceable, 5: steal something unique)
*Adds CE magboots