Properly handle stacks and filled items
Stacks should insert and eject properly now.  Additionally, code has been added to handle created items that have fill reagents, such as fire extinguishers and welding tools; these items can be created when there are no fill reagents in the lathe, and the resultant item will lose the fill reagent.  If there is some reagent in the lathe, it will be deducted from the stocks and the item will not be empty.  This removes the "lathe produces full items" bug from the old autolathe system.  It doesn't help with the protolathe creating full battery cells, though...