Adds support for filling air tanks and batteries.  Corrects disk uploading.
Properly recycles air tanks and builds air tanks with appropriate fill depending on stored gasses.  This involved creating reagents for CO2 and N2O gasses, which like the other maker reagents, are currently useless stub reagents.  Currently recycling tanks with unhandled gasses just dumps them into the square, they should really be captured in the future.

Disk uploads now function so long as the maker can handle all of the materials involved.  Disk-uploaded plans are persistent unless the machine is disassembled, and will not be wiped out if the machine is synched with a server (which doesn't happen yet anyway).  Disk uploaded plans are considered hacked plans, and the circuitboard must be hacked to upload them.  They are not currently saved to the circuitboard but should be later.

Currently the dump reagent button in the UI does not seem to work, I need to fix that.