Adds bottling reagents, UI tweaks.
Reagent bottling requires glass reagents in the lathe; without it, the buttons won't show up.  This is used with, for example, the biogenerator, which produces reagents like milk and fertilizer.  Bottled reagents can be produced in bulk as with material stacks; if you have enough glass and enough of the reagent, you can produce up to 10 bottles in a batch.  Materials that can be extruded into stacks cannot be bottled.

If the lathe can handle glass but has none, a disabled bottle button will show up as a placeholder to let the user know that glass should be inserted.  The bottle buttons, like the sheet extrusion buttons, have title text (mouse hover text) that is description, in order to save space.  Otherwise, the bottle and extrude buttons appear identical.

The main menu can be renamed on a machine-by-machine basis, allowing you to make the default category be named whatever you wish.  The default menu is set off graphically, so you can disable the main menu to make it appear flatter.  Note that if the main menu is hidden, anything added to the main menu will be unreachable.