Adds stock parts as item ingredients
Stock parts as ingredients currently behave mostly like reagents; they are not transferred into the resultant item, but simply lost in the construction process and created when an item containing them is recycled.  There really ought to be item procs associated with this, that may come later.

Stock parts are indexed in the maker_cost list by type, and subtypes work.  This means that requiring a basic manipulator will allow you to use medium or high quality manipulators, etc.  Note that the only checking at present is for subtypes, and not peers or quality (cells with X charge, etc).

At this point it probably needs improvement but I am running out of usable time...