Fixes seed extractor.  Faketurf bugfix.
The seed extractor no longer automatically bags seeds, and the existing tg interface has been restored.  Instead, you can vend seed bags from the extractor's interface.  These bags will be stamped with a short description of the seed's qualities, which is the same as the values in the extractor.

Faketurfs no longer runtime if they are garbage collected.  They will also be immune to explosions, which is fine since they are sort of artificial anyway.