Ankh Artifact and Various Tweaks
This adds in the Ankh artifact, which is a special type of artifact.
Only one can spawn per world, and it always has the same ability:
A_REVIVE.

The Ankh is an item that you can wear in your back slot. If you die
wearing it while it is activated, the Ankh will break and revive you.
This ability also prevents you from being gibbed by sources such as
bombs.

When you revive, you are fully healed and all defects are removed. You
also gain an item on your back called "Deo's Wings." This item is a pair
of wings. It can't be taken off, but you can use a button in the top
left of the screen to fold/unfold them.

While the wings are unfolded, you gain extra speed (2), the ability to
fly over tables, and they also function like a jetpack in space, leaving
behind a trail of golden sparkles. However, you cannot wear another pack
item (such as a backpack) until you fold them again.

The Ankh also gives you temporary (4 minute) protection against the cold
and general airlessness of space, which should be enough time to allow
you to fly back to the station.

This commit also gives several artifact types in-hand sprites for when
you're carrying them around. It also fixes a few misc bugs with
artifacts, and adds sprites for when you're wearing the armor artifact.
I also added an admin command that randomizes every artifact in the
game, for giggles.