Species Update Part 3: Get Horny
- Alien species can now have a second layer that's unaffected by your
custom alien color. This allows more details to be added to the sprites
without worrying about them being colored weird. In the case of
lizardpeople, this means that their claws will always be white.

- Lizardpeople sprites have been increased in brightness, and now have
more contrast. Basically, they look way better than they did before.

- Added a system that allows you to choose styles for species that do
not have hair. In the case of lizardpeople, this means you can choose
horn styles! At the moment, there are 12 styles in the game, but I'll
probably add more in the future.

- Increased max character slots to 8. Yes, even if you're not a donator.
Sorry, BYOND. (Hopefully this won't break saves, but it'd be a good idea
to back up the files in data/player_saves just in case)

- Started working on birdpeople, but they're not ready to be implemented
yet.

- Removes lizard titties (important)