Radscanner Fixes
This fixes the radscanner turning invisible when you use it, and also
fixes the interface not updating when you change the settings. Also,
increasing the intensity now also increases the wavelength slightly
more, meaning that the more powerful the radiation is the longer it
takes for the effect to kick in.